refactor: move TextField to src/v2/components
