+++ src/domains/users/Auth/components/LoginPage/LoginPage.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 
 import { Button, Checkbox, Form, FormLayout, Page } from "@shopify/polaris";
 import { routerPush } from "@Utils/router";
-import { TextField } from "../Login/TextField";
+import { TextField } from "@ComponentsV2/components/TextField";
 
 const formSchema = z.object({
   email: z
